Moved interfaces to separate files
