+++ HW2/Аlех_Сhaganоff/HW2/HW2/IRandomProvider.cs
@@ -1,0 +1,10 @@
+﻿
+using MyType = System.Int32;
+
+namespace HW2
+{
+    public interface IRandomProvider
+    {
+        MyType[] rand(Settings settings);
+    }
+}
