+++ HW2/Аlех_Сhaganоff/HW2/HW2/IReadInputProvider.cs
@@ -1,0 +1,7 @@
+﻿namespace HW2
+{
+    public interface IReadInputProvider
+    {
+        string readInput();
+    }
+}
