+++ HW2/Аlех_Сhaganоff/HW2/HW2/ISendOutputProvider.cs
@@ -1,0 +1,11 @@
+﻿
+using MyType = System.Int32;
+
+namespace HW2
+{
+    public interface ISendOutputProvider
+    {
+        void sendOutput(MyType[] currentNumbers);
+        void printOutput(string text);
+    }
+}
