+++ HW2/Аlех_Сhaganоff/HW2/HW2/ISettingsProvider.cs
@@ -1,0 +1,9 @@
+﻿
+namespace HW2
+{
+
+    public interface ISettingsProvider
+    {
+        Settings getSettings();
+    }
+}
