+++ HW2/Аlех_Сhaganоff/HW2/HW2/IStorageProvider.cs
@@ -1,0 +1,13 @@
+﻿
+using MyType = System.Int32;
+
+namespace HW2
+{
+    public interface IStorageProvider
+    {
+        MyType peek();
+        void push(MyType value);
+        MyType pop();
+        int Count { get; }
+    }
+}
