+++ HW2/Аlех_Сhaganоff/HW2/HW2/ITextMessagesProvider.cs
@@ -1,0 +1,7 @@
+﻿namespace HW2
+{
+    public interface ITextMessagesProvider
+    {
+        TextMessages getTextMessages();
+    }
+}
