+++ HW2/Аlех_Сhaganоff/HW2/HW2/TextMessages.cs
@@ -4,8 +4,6 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-using MyType = System.Int32;
-
 namespace HW2
 {
     public class TextMessages
